Remove run.mli and arrange.mli.
Run.ml has lots of useful features for the WASM frontend. Unfortunately, these are hidden behind a seriously limiting API. Same goes for Arrange.mli.

Really, it would be better to repackage the API. But for now, getting rid of the MLIs allows me to access the internal functions I need.